Profile view now has score username and an image
